Sistema/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg[binary]>=3.1
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
django-widget-tweaks>=1.5
openpyxl>=3.1
cloudinary>=1.40
django-cloudinary-storage>=0.3
django-simple-history>=3.7.0
twilio>=9.0
xhtml2pdf>=0.2.14
